requirements.txt
attrs==25.3.0
autocommand==2.2.2
build==1.2.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
cssselect==1.2.0
dbus-python==1.4.0
editables==0.5
faiss-cpu==1.11.0
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
gps==3.26.1
greenlet==3.2.3
hatchling==1.27.0
hf-xet==1.1.5
huggingface-hub==0.33.2
hyprshade==3.2.1
i3ipc==2.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
installer==0.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
joblib==1.5.1
lark==1.2.2
lxml==6.0.0
Mako==1.3.10.dev0
Markdown==3.8.2
MarkupSafe==3.0.2
mercurial==7.0.2
meson==1.8.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nwg-displays==0.3.25
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
poetry-core==2.1.3
pycairo==1.28.0
pycparser==2.22
PyGObject==3.52.3
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2025.2
pyxdg==0.28
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
screeninfo==0.8.1
sentence-transformers==5.0.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
trove-classifiers==2025.5.9.12
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waypaper==2.6
wheel==0.45.1
yt-dlp==2025.6.30
